rename file to suit content
