+++ examples/index.html
@@ -64,8 +64,8 @@
     <textarea cols="50" rows="10">The examples can be found here.</textarea>
 
     <script type="module">
-      // import {subscribe} from '../dist/index.esm.js'
-      import {subscribe} from 'https://unpkg.com/@github/paste-markdown/dist/index.esm.js'
+      import {subscribe} from '../dist/index.esm.js'
+      // import {subscribe} from 'https://unpkg.com/@github/paste-markdown/dist/index.esm.js'
       subscribe(document.querySelector('textarea'))
     </script>
   </body>
