+++ src/paste-markdown-table.ts
@@ -1,5 +1,5 @@
 import {insertText} from './text'
-import {isUnformatted} from './handlers'
+import {isUnformatted} from './paste-keyboard-shortcut-helper'
 
 export function install(el: HTMLElement): void {
   el.addEventListener('dragover', onDragover)
